LPS-153676 Move class to layout-test-util module
