+++ modules/apps/layout/layout-test-util/src/main/java/com/liferay/layout/test/util/LayoutContentPageEditorTestUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.content.page.editor.web.internal;
+package com.liferay.layout.test.util;
 
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.model.Company;
